wip: setting up wire for DI
