+++ example/main.go
@@ -4,11 +4,13 @@ import (
 	"net/http"
 
 	"github.com/roger-king/tasker"
+	tm "github.com/roger-king/tasker/models"
 )
 
 func main() {
-	t := tasker.New(&tasker.TaskerConfig{
-		MongoConnectionURL: "mongodb://appuser:appuser@localhost:27017/tasker?authSource=admin",
+	t := tasker.New(tm.TaskerConfig{
+		Type:               "mongo",
+		MongoConnectionURL: tm.MongoConnectionURL("mongodb://appuser:appuser@localhost:27017/tasker?authSource=admin"),
 	})
 	router := t.Start()
