+++ models/Config.go
@@ -1,0 +1,9 @@
+package models
+
+type MongoConnectionURL string
+type DBType string
+
+type TaskerConfig struct {
+	Type               DBType             `required:"true"`
+	MongoConnectionURL MongoConnectionURL `required:"true"`
+}
