+++ services/set.go
@@ -2,4 +2,4 @@ package services
 
 import "github.com/google/wire"
 
-var ServiceSet = wire.NewSet()
+var ServiceSet = wire.NewSet(NewMongoService, NewMongoConnection)
