+++ tasker.go
@@ -11,25 +11,18 @@ import (
 	_ "github.com/joho/godotenv/autoload"
 	cron "github.com/robfig/cron/v3"
 	"github.com/roger-king/tasker/handlers"
+	"github.com/roger-king/tasker/models"
 	"github.com/roger-king/tasker/services"
 	log "github.com/sirupsen/logrus"
 	"go.mongodb.org/mongo-driver/mongo"
 )
 
-type MongoConnectionURL string
-type DBType string
-
 // Tasker -
 type Tasker struct {
-	Config    TaskerConfig
+	Config    models.TaskerConfig
 	DB        *mongo.Client
 	Scheduler *cron.Cron
 	Router    *mux.Router
-}
-
-type TaskerConfig struct {
-	Type               DBType `required:"true"`
-	MongoConnectionURL MongoConnectionURL `required:"true"`
 }
 
 func init() {
@@ -46,15 +39,15 @@ func init() {
 
 var TaskerSet = wire.NewSet(services.ServiceSet, handlers.RouterSet, ProivdeTasker)
 
-func ProivdeTasker(tc TaskerConfig, r *mux.Router) *Tasker {
-	return &Tasker {
+func ProivdeTasker(tc models.TaskerConfig, r *mux.Router) *Tasker {
+	return &Tasker{
 		Config: tc,
 		Router: r,
 	}
 }
 
 // New - Creates a new instance of tasker
-func New(tc TaskerConfig) *Tasker {
+func New(tc models.TaskerConfig) *Tasker {
 	wire.Build(TaskerSet)
 	return nil
 }
